rename and fix package
