+++ parse/subgroups.go
@@ -1,6 +1,6 @@
-package Parse
+package parse
 
-import . "Schedule/Structure"
+import . "schedule/structure"
 
 func SubGroupParse(subject, typeOfLesson, teacherName, cabinet, dayOfWeek, numberLesson, week string) []Lesson {
