+++ structure/schedule_struct.go
@@ -1,4 +1,4 @@
-package Structure
+package structure
 
 import "strings"
